moved tests of child_process calls to a test file that renders more of the app - in order to test more expectations.
