+++ tests/app.test.js
@@ -1,18 +1,20 @@
 import path from 'path';
 import fs from 'fs';
+import events from 'events';
+import os from 'os';
+import { spawn } from 'child_process';
+jest.mock('child_process');
+import Stream from 'stream';
+
 import React from 'react';
-import { remote } from 'electron';
-import { fireEvent, render, waitFor } from '@testing-library/react'
+import { fireEvent, render, waitFor, within } from '@testing-library/react'
 import '@testing-library/jest-dom'
 
 import App from '../src/app';
-import { getInvestList, getFlaskIsReady } from '../src/server_requests';
+import { getInvestList, getSpec, fetchValidation } from '../src/server_requests';
 jest.mock('../src/server_requests');
 import { fileRegistry } from '../src/constants';
 import { cleanupDir } from '../src/utils'
-
-getFlaskIsReady.mockResolvedValue('Flask ready');
-getInvestList.mockResolvedValue({});
 
 afterAll(() => {
     cleanupDir(fileRegistry.TEMP_DIR)
@@ -20,6 +22,7 @@ afterAll(() => {
 })
 
 test('Recent Sessions: each has a button', async () => {
+  getInvestList.mockResolvedValue({});
   // This is a special json containing data used for testing
   const testJobsDatabase = path.join(__dirname, './data/jobdb.json');
   const { getByText, getByLabelText, debug } = render(
@@ -35,11 +38,11 @@ test('Recent Sessions: each has a button', async () => {
 })
 
 test('Settings dialog interactions: logging level', async () => {
-  
+  getInvestList.mockResolvedValue({});
   const DEFAULT = 'INFO';
 
   const { getByText, getByLabelText, debug } = render(
-    <App appdata={fileRegistry.JOBS_DATABASE}/>);
+    <App jobDatabase={fileRegistry.JOBS_DATABASE}/>);
 
   // Check the default settings
   fireEvent.click(getByText('Settings'));
@@ -64,13 +67,14 @@ test('Settings dialog interactions: logging level', async () => {
 })
 
 test('Settings dialog interactions: n workers', async () => {
+  getInvestList.mockResolvedValue({});
   const defaultValue = '-1';
   const newValue = '2'
   const badValue = 'a'
   const labelText = 'Taskgraph n_workers parameter'
 
   const { getByText, getByLabelText, debug } = render(
-    <App appdata={fileRegistry.JOBS_DATABASE}/>);
+    <App jobDatabase={fileRegistry.JOBS_DATABASE}/>);
 
   fireEvent.click(getByText('Settings'));
   const input = getByLabelText(labelText, { exact: false })
@@ -100,4 +104,115 @@ test('Settings dialog interactions: n workers', async () => {
   fireEvent.change(input, { target: { value: badValue } })
   expect(input.classList.contains('is-invalid')).toBeTruthy();
   expect(getByText('Save Changes')).toBeDisabled();
+})
+
+describe('InVEST subprocess testing', () => {
+  const spec = { args: {
+    workspace_dir: { 
+      name: 'Workspace', 
+      type: 'directory'},
+    results_suffix: {
+      name: 'Suffix',
+      type: 'freestyle_string'} } }
+
+  const dummyTextToLog = JSON.stringify(spec.args)
+  const logfileName = 'InVEST-natcap.invest.model-log-9999-99-99--99_99_99.txt'
+  const fakeWorkspace = 'tests/data'
+  const logfilePath = path.join(fakeWorkspace, logfileName)
+
+  const mockInvestProc = new events.EventEmitter();
+
+  beforeEach(() => {
+    // Need to reset these streams since mockInvestProc is shared by tests
+    // and the streams apparently receive the EOF signal in each test.
+    mockInvestProc.stdout = new Stream.Readable({
+      read() {},
+    })
+    mockInvestProc.stderr = new Stream.Readable({
+      read() {}
+    })
+    getSpec.mockResolvedValue(spec);
+    fetchValidation.mockResolvedValue([]);
+    getInvestList.mockResolvedValue(
+      {Carbon: {internal_name: 'carbon'}});
+    
+    spawn.mockImplementation((exe, cmdArgs, options) => {
+      // To simulate an invest model run, write a logfile to the workspace
+      // The line-ending is critical; the log is read with `tail.on('line'...)`
+      fs.writeFileSync(logfilePath, dummyTextToLog + os.EOL)
+      return mockInvestProc
+    })
+  })
+
+  afterEach(() => {
+    fs.unlinkSync(logfilePath)
+    jest.resetAllMocks();
+  })
+  
+  test('Invest subprocess - exit without error', async () => {
+    const { getByText, getByLabelText, ...utils } = render(
+      <App jobDatabase={fileRegistry.JOBS_DATABASE}/>);
+
+    const carbon = await utils.findByText('Carbon');
+    fireEvent.click(carbon);
+    const workspaceInput = await utils.findByLabelText(
+      RegExp(`${spec.args.workspace_dir.name}`))
+    fireEvent.change(workspaceInput, { target: { value: fakeWorkspace } })
+    
+    const execute = await utils.findByText('Execute');
+    fireEvent.click(execute);
+    
+    // Emit some stdout, pause, then exit without error
+    mockInvestProc.stdout.push('hello from stdout')
+    await new Promise(resolve => setTimeout(resolve, 2000))
+    mockInvestProc.emit('close', 0)  // 0 - exit w/o error
+    
+    await waitFor(() => {
+      expect(getByText('Log').classList.contains('active')).toBeTruthy();
+      // some text from the logfile should be rendered:
+      expect(getByText(dummyTextToLog, { exact: false }))
+        .toBeInTheDocument();
+      expect(getByText('Model Completed')).toBeInTheDocument();
+    });
+    // A recent session card should be rendered
+    const { findByText } = within(getByLabelText('Recent Sessions:'))
+    const cardText = await findByText(`${path.resolve(fakeWorkspace)}`)
+    expect(cardText).toBeInTheDocument()
+  })
+
+  test('Invest subprocess - exit with error', async () => {
+    // const { getByText, getByLabelText, utils } = renderInvestJob()
+    const { getByText, getByLabelText, ...utils } = render(
+      <App jobDatabase={fileRegistry.JOBS_DATABASE}/>);
+
+    const carbon = await utils.findByText('Carbon');
+    fireEvent.click(carbon);
+    const workspaceInput = await utils.findByLabelText(
+      RegExp(`${spec.args.workspace_dir.name}`))
+    fireEvent.change(workspaceInput, { target: { value: fakeWorkspace } })
+    
+    const execute = await utils.findByText('Execute');
+    fireEvent.click(execute);
+    
+    const errorMessage = 'fail'
+    // Emit some stdout, some stderr, then pause and exit with error
+    mockInvestProc.stdout.push('hello from stdout')
+    mockInvestProc.stderr.push(errorMessage)
+    await new Promise(resolve => setTimeout(resolve, 2000))
+    mockInvestProc.emit('close', 1)  // 1 - exit w/ error
+    
+    await waitFor(() => {
+      expect(getByText('Log').classList.contains('active')).toBeTruthy();
+      // some text from the logfile should be rendered:
+      expect(getByText(dummyTextToLog, { exact: false }))
+        .toBeInTheDocument();
+      // stderr text should be rendered 
+      expect(getByText(errorMessage)).toHaveClass('alert-danger');
+    });
+    // A recent session card should be rendered
+    const { findByText } = within(getByLabelText('Recent Sessions:'))
+    const cardText = await findByText(`${path.resolve(fakeWorkspace)}`)
+    expect(cardText).toBeInTheDocument()
+  })
+
 })
